+++ internal/plugin/plugin_manifest_parser.go
@@ -1,0 +1,56 @@
+// Copyright 2024 MongoDB Inc
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package plugin
+
+import (
+	"fmt"
+	"os"
+	"path/filepath"
+
+	"gopkg.in/yaml.v2"
+)
+
+func parseManifestFile(manifestFileData []byte) (*PluginManifest, error) {
+	var pluginManifest PluginManifest
+
+	if err := yaml.Unmarshal(manifestFileData, &pluginManifest); err != nil {
+		return nil, err
+	}
+
+	return &pluginManifest, nil
+}
+
+func getManifestFileBytes(pluginDirectoryPath string) ([]byte, error) {
+	validManifestFilenames := []string{"manifest.yml", "manifest.yaml"}
+
+	for _, filename := range validManifestFilenames {
+		manifestFilePath := filepath.Join(pluginDirectoryPath, filename)
+
+		info, err := os.Stat(manifestFilePath)
+		if os.IsNotExist(err) || info.IsDir() {
+			continue
+		}
+
+		manifestFileData, err := os.ReadFile(manifestFilePath)
+
+		if err != nil {
+			continue
+		}
+
+		return manifestFileData, nil
+	}
+
+	return nil, fmt.Errorf("plugin invalid: manifest file does not exist in plugin folder %s", pluginDirectoryPath)
+}
